fix(rocm): address second review — remove HSA_OVERRIDE, add context floor, use actual cache dtype
- Remove automatic HSA_OVERRIDE_GFX_VERSION injection entirely; the
  detected real architecture is a no-op when ROCm supports the GPU and
  useless when it does not. Users who need override can set the env var
  manually, matching the docker/rocm/README.md guidance.
- Add 2048-token lower bound to resolveNumCtxVRAMAware; when free VRAM
  is tight and the ceiling falls below 2048, log the reason and fall
  back to the existing ResolveNumCtx logic instead of starting with an
  unusably small context.
- Resolve cache_type_k/cache_type_v before calling resolveNumCtxVRAMAware
  and pass the actual dtype size to KVCacheBytesPerToken instead of
  hardcoding f16 (2). A q8_0 cache now correctly estimates half the KV
  cost per token.
- Remove redundant min() around capNumCtxToModelMax, which already
  returns at most its argument.
- Add doc comment noting the multi-engine VRAM limitation: free VRAM
  is sampled at load time, so the first model sizes to all available
  VRAM.

+++ internal/ggufmeta/reader.go
@@ -245,7 +245,7 @@ func (m *Metadata) IsVisionProjector() bool {
 // KVCacheBytesPerToken estimates the KV cache bytes per token from GGUF
 // metadata. Returns 0 if the required fields are missing.
 // dtypeSize is the bytes per element of one KV entry (e.g. 2 for f16).
-func KVCacheBytesPerToken(path string, dtypeSize int) int64 {
+func KVCacheBytesPerToken(path string, dtypeSize float64) int64 {
 	meta, err := ReadFile(path)
 	if err != nil {
 		return 0
@@ -287,7 +287,7 @@ func KVCacheBytesPerToken(path string, dtypeSize int) int64 {
 	if dtypeSize <= 0 {
 		dtypeSize = 2
 	}
-	return int64(blockCount) * int64(headCountKV) * int64(keyLength+valueLength) * int64(dtypeSize)
+	return int64(float64(blockCount) * float64(headCountKV) * float64(keyLength+valueLength) * dtypeSize)
 }
 
 func readValue(reader io.Reader, valueType uint32) (value, bool, error) {
